pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentb"
version = "0.1.0"
description = "Reflexive AI agent for browser automation tasks"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Agent B Team"}
]
dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "chromadb>=0.4.20",
    "sentence-transformers>=2.3.0",
    "anthropic>=0.14.0",
    "Pillow>=10.1.0",
    "tenacity>=8.2.3",
    "structlog>=24.1.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
]

[project.scripts]
agentb = "agentb.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
